Create subdivide crossing_counter and label_export
To improve model performance, we will subdivide
our images into a 4x4 grid and label each of the
16 squares within the grid.

We currently have the grid properly overlaid on
the satellite image. Next up, we must keep track
of the count of crossings falling within each
square. This info will be pushed to
subdivide_label_export, which will then produce
an image for that square with the correct label.