Change and update many files
Coordinates finalized (resulting in asia_final
and euro_final)

A lot of small scripts were made and modified for
the sake of dataset generations (both images and
coordinates). With our dataset assembled, these
scripts are no longer relevant, but it would be
nice to go back and clean these up. For instance,
we probably don't need three separate variations
of crossing_counter...

Some changes were made to model, but many more are
on the way. Need to get batch loading working to
deal with large dataset...